Renamed ddl and dml
