+++ src/main/java/tech/ydb/io/r2dbc/query/YdbQuery.java
@@ -56,8 +56,6 @@ public class YdbQuery {
                             .append(" AS ")
                             .append(prmType)
                             .append(";\n");
-
-
                 }
             } else if (!indexesArgsNames.isEmpty()) {
                 yql.append("-- DECLARE ").append(indexesArgsNames.size()).append(" PARAMETERS").append("\n");
