+++ src/main/java/tech/ydb/io/r2dbc/statement/YdbDMLStatement.java
@@ -16,10 +16,12 @@
 
 package tech.ydb.io.r2dbc.statement;
 
+import java.sql.SQLException;
 
 import io.r2dbc.spi.Result;
 import io.r2dbc.spi.Statement;
 import org.reactivestreams.Publisher;
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import tech.ydb.io.r2dbc.query.YdbQuery;
 import tech.ydb.io.r2dbc.state.YdbConnectionState;
@@ -27,39 +29,43 @@ import tech.ydb.io.r2dbc.state.YdbConnectionState;
 /**
  * @author Egor Kuleshov
  */
-public class DDLYdbStatement extends YdbStatement {
-    public DDLYdbStatement(YdbQuery query, YdbConnectionState ydbConnectionState) {
+public class YdbDMLStatement extends YdbStatement {
+    public YdbDMLStatement(YdbQuery query, YdbConnectionState ydbConnectionState) {
         super(query, ydbConnectionState);
-    }
-    @Override
-    public Statement add() {
-        return this;
     }
 
     @Override
     public Statement bind(int i, Object o) {
-        return this;
+        return null;
     }
 
     @Override
     public Statement bind(String s, Object o) {
-        return this;
+        return null;
     }
 
     @Override
     public Statement bindNull(int i, Class<?> aClass) {
-        return this;
+        return null;
     }
 
     @Override
     public Statement bindNull(String s, Class<?> aClass) {
-        return this;
+        return null;
     }
 
     @Override
     public Publisher<? extends Result> execute() {
         try {
-            return connectionState.executeSchemaQuery(query.getYqlQuery(bindings.getCurrent()), bindings.getCurrent());
+            return Flux.fromIterable(bindings)
+                    .flatMap(binding -> {
+                        try {
+                            return connectionState.executeDataQuery(query.getYqlQuery(bindings.getCurrent()),
+                                            bindings.getCurrent());
+                        } catch (SQLException e) {
+                            return Mono.error(e);
+                        }
+                    });
         } catch (Exception e) {
             return Mono.error(e);
         }
